+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Add Curse packs from the packs tab
 - [#370] fix JVM arguments not adding in correct order
 - Fix adding mods showing modpacks and add dropdown for resource packs
+- Fix changing pages when adding mods not keeping search query